Use size_t instead of int everywhere

+++ src/graph.c
@@ -17,12 +17,12 @@ Graph *createGraph(size_t nodesAmount) {
     return graph;
 }
 
-void addLine(Graph *graph, int sourceIndex, int destinationIndex) {
+void addLine(Graph *graph, size_t sourceIndex, size_t destinationIndex) {
     graph->adjacencyLists[sourceIndex]->head->edgesCount++;
     addToList(graph->adjacencyLists[sourceIndex], destinationIndex);
 }
 
-void tryToVisitAllNodes(Graph *graph, int startIndex, int *visitedNodes, size_t iteration) {
+void tryToVisitAllNodes(Graph *graph, size_t startIndex, size_t *visitedNodes, size_t iteration) {
     if (visitedNodes[startIndex] == 1) {
         return;
     }
@@ -37,7 +37,7 @@ void tryToVisitAllNodes(Graph *graph, int startIndex, int *visitedNodes, size_t 
     }
 }
 
-void getLongestCycle(Graph *graph, int currentIndex, int *visited, int startNodeData, List **longestCycle,
+void getLongestCycle(Graph *graph, size_t currentIndex, size_t *visited, size_t startNodeData, List **longestCycle,
                      size_t *cyclesAmount) {
     if (currentIndex == graph->size) {
         return;
@@ -78,7 +78,7 @@ List *getLongestGraphCycle(Graph *graph) {
     List *cycles[2] = {createList(), createList()};
     size_t cyclesAmount = 0;
     for (size_t i = 0; i < graph->size; i++) {
-        int visited[graph->size];
+        size_t visited[graph->size];
         memset(visited, 0, sizeof(visited));
         getLongestCycle(graph, i, visited, i, cycles, &cyclesAmount);
     }
@@ -90,8 +90,8 @@ List *getLongestGraphCycle(Graph *graph) {
     return cycles[0];
 }
 
-int isGraphConnected(Graph *graph) {
-    int visitedNodes[graph->size];
+size_t isGraphConnected(Graph *graph) {
+    size_t visitedNodes[graph->size];
     memset(visitedNodes, 0, sizeof(visitedNodes));
     for (size_t i = 0; i < graph->size; i++) {
         tryToVisitAllNodes(graph, i, visitedNodes, i);
